requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
PyJWT==1.7.1
pymongo==3.12.3
pytz==2023.3
sqlparse==0.2.4

